rename switch to switchButton
