+++ src/sentry/static/sentry/app/components/switchButton.tsx
