+++ src/sentry/static/sentry/app/views/organizationIntegrations/installedPlugin.tsx
@@ -12,7 +12,7 @@ import Alert from 'app/components/alert';
 import Button from 'app/components/button';
 import Confirm from 'app/components/confirm';
 import ProjectBadge from 'app/components/idBadge/projectBadge';
-import Switch from 'app/components/switch';
+import Switch from 'app/components/switchButton';
 import {IconDelete, IconFlag, IconSettings} from 'app/icons';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
