+++ src/sentry/static/sentry/app/views/settings/project/projectFilters/projectFiltersSettings.tsx
@@ -13,7 +13,7 @@ import {
   PanelHeader,
   PanelItem,
 } from 'app/components/panels';
-import Switch from 'app/components/switch';
+import Switch from 'app/components/switchButton';
 import filterGroups, {customFilterFields} from 'app/data/forms/inboundFilters';
 import {t} from 'app/locale';
 import HookStore from 'app/stores/hookStore';
