+++ src/sentry/static/sentry/app/views/settings/projectPlugins/projectPluginRow.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 
 import Access from 'app/components/acl/access';
 import ExternalLink from 'app/components/links/externalLink';
-import Switch from 'app/components/switch';
+import Switch from 'app/components/switchButton';
 import {t} from 'app/locale';
 import PluginIcon from 'app/plugins/components/pluginIcon';
 import {Organization, Plugin, Project} from 'app/types';
